Refactor Action hierarchy, including a prototype of its implementation
Add new package to contain action hierarchy, add new actions to actually make all possible game actions and implements logic about if action is valid or not. Implement part of the constructors and perform methods, but implementation is not finished because model lacks weapon classes and grab features in board.