Add new tests, change some exception from checked to unchecked
Some exception in the project were checked, but they could be avoided with proper control methods. It was useless to use a try catch approach to define logic decision flow, it only led to uglier code. Now this exceptions are RuntimeException and are still used in code where they were before this commit, they can just be avoided with proper methods without using try catch everytime.