Add generic selection checking
Now all selections (target, position, etc...) are checked before they
are used. An error is signaled to the view and the selection is
requested again if it is not deemed valid.

Also add test (testing plasma gun) to showcase an invalid selection. A
mock is used to verify that the View.reportError() method is called.